requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
tqdm>=4.65
openai>=1.0
python-dotenv>=1.0
requests>=2.31
jupyter>=1.0
ipykernel>=6.0